# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SubsMarket API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": [
          "identity"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_me_get",
        "parameters": [
          {
            "name": "x-telegram-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          },
          {
            "name": "x-dev-telegram-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Telegram-User-Id"
            }
          },
          {
            "name": "x-dev-telegram-username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Telegram-Username"
            }
          },
          {
            "name": "x-dev-telegram-first-name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Telegram-First-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/refresh-telegram-profile": {
      "patch": {
        "tags": [
          "identity"
        ],
        "summary": "Refresh Telegram Profile",
        "operationId": "refresh_telegram_profile_api_me_refresh_telegram_profile_patch",
        "parameters": [
          {
            "name": "x-telegram-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Init-Data"
            }
          },
          {
            "name": "x-dev-telegram-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Telegram-User-Id"
            }
          },
          {
            "name": "x-dev-telegram-username",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Telegram-Username"
            }
          },
          {
            "name": "x-dev-telegram-first-name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Telegram-First-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/internal/jobs/run-due": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Post Run Due Jobs",
        "operationId": "post_run_due_jobs_api_internal_jobs_run_due_post",
        "parameters": [
          {
            "name": "x-internal-job-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Job-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDueJobsResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/internal/jobs/dispatch-notifications": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Post Dispatch Notifications",
        "operationId": "post_dispatch_notifications_api_internal_jobs_dispatch_notifications_post",
        "parameters": [
          {
            "name": "x-internal-job-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Job-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchNotificationsResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/internal/jobs/status": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Internal Jobs Status",
        "operationId": "get_internal_jobs_status_api_internal_jobs_status_get",
        "parameters": [
          {
            "name": "x-internal-job-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Job-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsStatusResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DispatchNotificationsResult": {
        "properties": {
          "selected": {
            "type": "integer",
            "title": "Selected"
          },
          "sent": {
            "type": "integer",
            "title": "Sent"
          },
          "retried": {
            "type": "integer",
            "title": "Retried"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          }
        },
        "type": "object",
        "required": [
          "selected",
          "sent",
          "retried",
          "failed"
        ],
        "title": "DispatchNotificationsResult"
      },
      "DueBacklogStatus": {
        "properties": {
          "expired_family_requests": {
            "type": "integer",
            "title": "Expired Family Requests"
          },
          "access_confirmations_overdue": {
            "type": "integer",
            "title": "Access Confirmations Overdue"
          },
          "first_payments_overdue": {
            "type": "integer",
            "title": "First Payments Overdue"
          },
          "regular_payment_creation_due": {
            "type": "integer",
            "title": "Regular Payment Creation Due"
          },
          "regular_payments_to_activate": {
            "type": "integer",
            "title": "Regular Payments To Activate"
          },
          "regular_payments_overdue": {
            "type": "integer",
            "title": "Regular Payments Overdue"
          },
          "owner_payment_confirmations_waiting": {
            "type": "integer",
            "title": "Owner Payment Confirmations Waiting"
          },
          "closing_acknowledgements_due": {
            "type": "integer",
            "title": "Closing Acknowledgements Due"
          },
          "member_removals_due": {
            "type": "integer",
            "title": "Member Removals Due"
          },
          "family_closures_due": {
            "type": "integer",
            "title": "Family Closures Due"
          },
          "per_step_capacity": {
            "type": "integer",
            "title": "Per Step Capacity"
          }
        },
        "type": "object",
        "required": [
          "expired_family_requests",
          "access_confirmations_overdue",
          "first_payments_overdue",
          "regular_payment_creation_due",
          "regular_payments_to_activate",
          "regular_payments_overdue",
          "owner_payment_confirmations_waiting",
          "closing_acknowledgements_due",
          "member_removals_due",
          "family_closures_due",
          "per_step_capacity"
        ],
        "title": "DueBacklogStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobsStatusResult": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "notification_queue": {
            "$ref": "#/components/schemas/NotificationQueueStatus"
          },
          "due_backlog": {
            "$ref": "#/components/schemas/DueBacklogStatus"
          },
          "recent_notification_failures": {
            "items": {
              "$ref": "#/components/schemas/NotificationFailureSample"
            },
            "type": "array",
            "title": "Recent Notification Failures"
          }
        },
        "type": "object",
        "required": [
          "status",
          "checked_at",
          "notification_queue",
          "due_backlog"
        ],
        "title": "JobsStatusResult"
      },
      "MeResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "MeResponse"
      },
...
      "NotificationFailureSample": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "failed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "attempts"
        ],
        "title": "NotificationFailureSample"
      },
      "NotificationQueueStatus": {
        "properties": {
          "pending_total": {
            "type": "integer",
            "title": "Pending Total"
          },
          "pending_due": {
            "type": "integer",
            "title": "Pending Due"
          },
          "pending_future": {
            "type": "integer",
            "title": "Pending Future"
          },
          "sent_total": {
            "type": "integer",
            "title": "Sent Total"
          },
          "failed_total": {
            "type": "integer",
            "title": "Failed Total"
          },
          "cancelled_total": {
            "type": "integer",
            "title": "Cancelled Total"
          },
          "stale_due": {
            "type": "integer",
            "title": "Stale Due"
          },
          "failed_last_24h": {
            "type": "integer",
            "title": "Failed Last 24H"
          },
          "oldest_due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Due At"
          },
          "oldest_pending_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Pending At"
          },
          "dispatch_capacity_per_run": {
            "type": "integer",
            "title": "Dispatch Capacity Per Run"
          }
        },
        "type": "object",
        "required": [
          "pending_total",
          "pending_due",
          "pending_future",
          "sent_total",
          "failed_total",
          "cancelled_total",
          "stale_due",
          "failed_last_24h",
          "dispatch_capacity_per_run"
        ],
        "title": "NotificationQueueStatus"
      },
...
      "RunDueJobError": {
        "properties": {
          "step": {
            "type": "string",
            "title": "Step"
          },
          "error_type": {
            "type": "string",
            "title": "Error Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "step",
          "error_type",
          "message"
        ],
        "title": "RunDueJobError"
      },
      "RunDueJobsResult": {
        "properties": {
          "expired_family_requests": {
            "type": "integer",
            "title": "Expired Family Requests"
          },
          "access_confirmation_reminders_sent": {
            "type": "integer",
            "title": "Access Confirmation Reminders Sent"
          },
          "overdue_first_payments": {
            "type": "integer",
            "title": "Overdue First Payments"
          },
          "created_regular_payments": {
            "type": "integer",
            "title": "Created Regular Payments"
          },
          "activated_regular_payments": {
            "type": "integer",
            "title": "Activated Regular Payments"
          },
          "overdue_regular_payments": {
            "type": "integer",
            "title": "Overdue Regular Payments"
          },
          "regular_payment_reminders_sent": {
            "type": "integer",
            "title": "Regular Payment Reminders Sent"
          },
          "owner_payment_confirmation_reminders_sent": {
            "type": "integer",
            "title": "Owner Payment Confirmation Reminders Sent"
          },
          "closing_acknowledgement_reminders_sent": {
            "type": "integer",
            "title": "Closing Acknowledgement Reminders Sent"
          },
          "executed_member_removals": {
            "type": "integer",
            "title": "Executed Member Removals"
          },
          "closed_families": {
            "type": "integer",
            "title": "Closed Families"
          },
          "notification_jobs_created": {
            "type": "integer",
            "title": "Notification Jobs Created"
          },
          "job_errors": {
            "items": {
              "$ref": "#/components/schemas/RunDueJobError"
            },
            "type": "array",
            "title": "Job Errors"
          }
        },
        "type": "object",
        "required": [
          "expired_family_requests",
          "access_confirmation_reminders_sent",
          "overdue_first_payments",
          "created_regular_payments",
          "activated_regular_payments",
          "overdue_regular_payments",
          "regular_payment_reminders_sent",
          "owner_payment_confirmation_reminders_sent",
          "closing_acknowledgement_reminders_sent",
          "executed_member_removals",
          "closed_families",
          "notification_jobs_created"
        ],
        "title": "RunDueJobsResult"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "telegram_user_id": {
            "type": "integer",
            "title": "Telegram User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "telegram_user_id",
          "username",
          "first_name",
          "last_name",
          "photo_url",
          "status"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}